+++ src/core/opengl/render_gl.h
@@ -755,7 +755,7 @@ public:
                             StdPicture &tx,
                             XTColor color = XTColor()) override;
 
-    void renderTexture(double xDst, double yDst, double wDst, double hDst,
+    void renderTexture(int xDst, int yDst, int wDst, int hDst,
                        StdPicture &tx,
                        int xSrc, int ySrc,
                        XTColor color = XTColor()) override;
@@ -766,7 +766,7 @@ public:
                          int16_t rotateAngle = 0, Point_t *center = nullptr, unsigned int flip = X_FLIP_NONE,
                          XTColor color = XTColor()) override;
 
-    void renderTexture(float xDst, float yDst, StdPicture &tx,
+    void renderTexture(int xDst, int yDst, StdPicture &tx,
                        XTColor color = XTColor()) override;
 
     void renderParticleSystem(StdPicture &tx,
